ISPN-13657 Upgrade jboss-marshalling to 2.0.12.Final
