+++ server/integration/testsuite/src/test/java/org/infinispan/server/test/client/hotrod/CustomPojoEventFilterFactory.java
@@ -7,7 +7,6 @@ import org.infinispan.metadata.Metadata;
 import org.infinispan.notifications.cachelistener.filter.CacheEventFilter;
 import org.infinispan.notifications.cachelistener.filter.CacheEventFilterFactory;
 import org.infinispan.notifications.cachelistener.filter.EventType;
-import org.infinispan.server.test.client.hotrod.AbstractRemoteCacheIT.Person;
 
 @NamedFactory(name = "pojo-filter-factory")
 public class CustomPojoEventFilterFactory implements CacheEventFilterFactory {
